ci(bithound): mute eslint-plugin-json issues
eslint-plugin-json uses shelljs, which is issued with a security advice of high severity, due to
possibility of command injection. Currently there is no way to fix this, and this package is only
used on development envs, so I will mute it until a fixed version of that package gets released.